+++ src/js/ModeSwitcher.js
@@ -2,64 +2,48 @@ var ContextMenu = require('./ContextMenu');
 
 /**
  * Create a select box to be used in the editor menu's, which allows to switch mode
- * @param {Object} editor
+ * @param {HTMLElement} container
  * @param {String[]} modes  Available modes: 'code', 'form', 'text', 'tree', 'view'
  * @param {String} current  Available modes: 'code', 'form', 'text', 'tree', 'view'
- * @returns {HTMLElement} box
+ * @param {function(mode: string)} onSwitch  Callback invoked on switch
+ * @constructor
  */
-function createModeSwitcher(editor, modes, current) {
-  // TODO: decouple mode switcher from editor
-
-  /**
-   * Switch the mode of the editor
-   * @param {String} mode
-   */
-  function switchMode(mode) {
-    // switch mode
-    editor.setMode(mode);
-
-    // restore focus on mode box
-    var modeBox = editor.dom && editor.dom.modeBox;
-    if (modeBox) {
-      modeBox.focus();
-    }
-  }
-
+function ModeSwitcher(container, modes, current, onSwitch) {
   // available modes
   var availableModes = {
     code: {
       'text': 'Code',
       'title': 'Switch to code highlighter',
       'click': function () {
-        switchMode('code')
+        onSwitch('code')
       }
     },
     form: {
       'text': 'Form',
       'title': 'Switch to form editor',
       'click': function () {
-        switchMode('form');
+        onSwitch('form');
       }
     },
     text: {
       'text': 'Text',
       'title': 'Switch to plain text editor',
       'click': function () {
-        switchMode('text');
+        onSwitch('text');
       }
     },
     tree: {
       'text': 'Tree',
       'title': 'Switch to tree editor',
       'click': function () {
-        switchMode('tree');
+        onSwitch('tree');
       }
     },
     view: {
       'text': 'View',
       'title': 'Switch to tree view',
       'click': function () {
-        switchMode('view');
+        onSwitch('view');
       }
     }
   };
@@ -94,12 +78,35 @@ function createModeSwitcher(editor, modes, current) {
     menu.show(box);
   };
 
-  var div = document.createElement('div');
-  div.className = 'jsoneditor-modes';
-  div.style.position = 'relative';
-  div.appendChild(box);
+  var frame = document.createElement('div');
+  frame.className = 'jsoneditor-modes';
+  frame.style.position = 'relative';
+  frame.appendChild(box);
 
-  return div;
+  container.appendChild(frame);
+
+  this.dom = {
+    container: container,
+    box: box,
+    frame: frame
+  };
 }
 
-exports.create = createModeSwitcher;
+/**
+ * Set focus to switcher
+ */
+ModeSwitcher.prototype.focus = function () {
+  this.dom.box.focus();
+};
+
+/**
+ * Destroy the ModeSwitcher, remove from DOM
+ */
+ModeSwitcher.prototype.destroy = function () {
+  if (this.dom && this.dom.frame && this.dom.frame.parentNode) {
+    this.dom.frame.parentNode.removeChild(this.dom.frame);
+  }
+  this.dom = null;
+};
+
+module.exports = ModeSwitcher;